Improved Output for missing / wrong configuration
The module will now notify the user in case no parent endpoints have been configured, which will cause the icinga2.conf to not change after a run.
In addition the module will throw an exception in case a global zone is configured as parent agent zone (which is not possible).